requirements.txt

python-dotenv
langchain-google-genai
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
langgraph 